AVG-396: course archive lock enforcing (#325)
* AVG-396: usercourse getter endpoint enhanced

* AVG-396: archived course list visualization

* AVG-396: Internal course controller and internal course getter

* AVG-396: Archived course moved from 403 to 410 Gone

* Better date comparison in areTimestampsValid

* Removed useless log from webapp

* Enforcing 410 Gone for archived courses

* Cannot join trial after deadline fix

* Cannot submit if course is archived fix

* AVG-396: ArchivedCourseError added

* AVG-396: cannot create trial if course is archived

* AVG-396: cannot run similarity tool if course is archived

* AVG-396: archived course handling in course detail

* AVG-396: archived course handling in exercise screen

* AVG-396: submit button hidden if course is archived

* AVG-396: create exercise hidden if course is archived

* fixed exercise list top margin in student trial view

* AVG-396: better archived course lock enforcing in submission editor

* AVG-396: archived course lock enforced in deadline stack

* AVG-396: archived course lock enforced in trial detail screen

* CourseDetailDto added to exerciseService

* CourseRole Lombok optimization

* internal getUserCourseDetail endpoint refactor

* getUserCourseRoleDetail refactored to return CourseDetailDto

codebase and tests adapted

* AVG-396: archived courses hidden in trial and exercise creation

* AVG-396: archived course lock extended to trial and exercise creation

* AVG-396: review fixes

+++ backend/services/exerciseservice/src/main/java/eu/mostserene/avogador/exerciseservice/testcases/TestcaseController.java
@@ -41,8 +41,8 @@ public class TestcaseController {
                 .orElseThrow(() -> new NotFoundException("Not found exercise with id: " + exerciseId));
         var trial = trialService.getTrialById(exercise.getTrial().getId())
                 .orElseThrow(() -> new NotFoundException("Not found trial with id: " + exercise.getTrial().getId()));
-        var courseRole = userCourseService.getUserCourseRole(trial.getCourseId(), user.getId())
-                .orElseThrow(() -> new ForbiddenException(user));
+        var courseRole = userCourseService.getUserCourseRoleDetail(trial.getCourseId(), user.getId())
+                .orElseThrow(() -> new ForbiddenException(user)).getRole();
 
         if (courseRole.getClearance() < CourseRole.STUDENT.getClearance() && !user.getIsSuperuser()) {
             throw new ForbiddenException(user);
@@ -65,8 +65,8 @@ public class TestcaseController {
                 .orElseThrow(() -> new NotFoundException("Not found exercise with id: " + exerciseId));
         var trial = trialService.getTrialById(exercise.getTrial().getId())
                 .orElseThrow(() -> new NotFoundException("Not found trial with id: " + exercise.getTrial().getId()));
-        var courseRole = userCourseService.getUserCourseRole(trial.getCourseId(), user.getId())
-                .orElseThrow(() -> new ForbiddenException(user));
+        var courseRole = userCourseService.getUserCourseRoleDetail(trial.getCourseId(), user.getId())
+                .orElseThrow(() -> new ForbiddenException(user)).getRole();
 
         if (courseRole.getClearance() < CourseRole.STUDENT.getClearance() && !user.getIsSuperuser()) {
             throw new ForbiddenException(user);
@@ -106,6 +106,13 @@ public class TestcaseController {
             @PathVariable UUID exerciseId,
             @RequestBody TestcaseDetailDto testcase) {
         var exercise = getExerciseIfCollaboratorClearance(exerciseId, user);
+        var courseDetail = userCourseService.getUserCourseRoleDetail(exercise.getTrial().getCourseId(), user.getId())
+                .orElseThrow(() -> new ForbiddenException(user));
+
+        if (courseDetail.getIsArchived()) {
+            throw new ResponseStatusException(HttpStatus.GONE, "This course has been archived");
+        }
+
         testcase.setExerciseId(exerciseId);
 
         if (testcase.getIndex() == null) {
@@ -123,6 +130,13 @@ public class TestcaseController {
     private void deleteTestcase(@RequestHeader(name = "User") UserDto user, @PathVariable UUID exerciseId, @PathVariable UUID testcaseId) {
         var exercise = getExerciseIfCollaboratorClearance(exerciseId, user);
 
+        var courseDetail = userCourseService.getUserCourseRoleDetail(exercise.getTrial().getCourseId(), user.getId())
+                .orElseThrow(() -> new ForbiddenException(user));
+
+        if (courseDetail.getIsArchived()) {
+            throw new ResponseStatusException(HttpStatus.GONE, "This course has been archived");
+        }
+
         testcaseService.deleteTestcase(exercise, testcaseId);
     }
 
@@ -131,8 +145,8 @@ public class TestcaseController {
                 .orElseThrow(() -> new NotFoundException("Not found exercise with id: " + exerciseId));
         var trial = trialService.getTrialById(exercise.getTrial().getId())
                 .orElseThrow(() -> new NotFoundException("Not found trial with id: " + exercise.getTrial().getId()));
-        var courseRole = userCourseService.getUserCourseRole(trial.getCourseId(), user.getId())
-                .orElseThrow(() -> new ForbiddenException(user));
+        var courseRole = userCourseService.getUserCourseRoleDetail(trial.getCourseId(), user.getId())
+                .orElseThrow(() -> new ForbiddenException(user)).getRole();
 
         if (courseRole.getClearance() < CourseRole.COLLABORATOR.getClearance() && !user.getIsSuperuser()) {
             throw new ForbiddenException(user);
